cut down on PID memory storage by half by using Shorts instead of Integers

+++ android-obd-library/src/main/java/com/obd/lib/models/PID.java
@@ -26,6 +26,6 @@ public class PID {
     public String Max;
     public String Units;
     public String Formula;
-    public Byte[] Data;
+    public Short[] Data;
     public String CalculatedResult;
 }
